refactor indentation logic to remove repeated code
