+++ org.rebecalang.afra.ideplugin/src/org/rebecalang/afra/ideplugin/editors/formatter/PropertyFormatter.java
@@ -1,0 +1,38 @@
+package org.rebecalang.afra.ideplugin.editors.formatter;
+
+import org.eclipse.jface.text.BadLocationException;
+import org.eclipse.jface.text.IDocument;
+import org.eclipse.jface.text.IRegion;
+
+/**
+ * Formatter for Property (.property) files that correctly handles comments
+ */
+public class PropertyFormatter implements IAfraFormatter {
+    
+    @Override
+    public String format(IDocument document) {
+        try {
+            String content = document.get();
+            return FormatterUtils.formatContent(content);
+        } catch (Exception e) {
+            e.printStackTrace();
+            return document.get();
+        }
+    }
+    
+    @Override
+    public String format(IDocument document, IRegion region) {
+        try {
+            String content = document.get(region.getOffset(), region.getLength());
+            return FormatterUtils.formatContent(content);
+        } catch (BadLocationException e) {
+            e.printStackTrace();
+            return document.get();
+        }
+    }
+    
+    @Override
+    public String getIndentString() {
+        return FormatterUtils.getIndentString();
+    }
+}
